# Remove check for ring information from Atom::Match